azure-vs-aws/azure-implement/requirements.txt
azure-ai-documentintelligence>=1.0.0
azure-core>=1.30.0
python-dotenv>=1.0.0

# new
azure-storage-blob
pydantic-settings